Use plones IContentIcon adapter when available.
Changes the helper to use Plone's IContentIcon helper instead of getIcon
directly.
This allows us to change the behavior by overwriting the adapters, for
example for replacing mimetype icons with font-icons or sprites.

This may not work on older Plone sites and it does not work when the
item is a Solr flair. In those cases we fallback to the old
implementation.