Fix concurrency bug in generator: data was leaking between threads.
When registering a utility as factory, the factory is called on ZCML
parse time and returns a singleton object which is kept.
When the generator utility was built, we relied on the factory beeing
called on utility lookup time, but this is not how it works.

Using the same object for multiple "generate" calls has side effects,
especially when used concurrently in multiple threads.

Since the generator utility is used in many places in our code base,
we decided to not change the public facing API but solve the problem
internally by using a real singleton object, which delegates
"generate" calls through freshly instantiated table generator objects.

We couldn't change the table generator's generate method to be static,
because it uses the view page template, which relies on object state.