CatalogTableSource: Exclude searchroot from results:
By default, the CatalogTableSource will now take care
to not include the searchroot itself in the results.

This fixes a bug where otherwise for listings on containers
the folderish context itself was included in the list of its
children.

This behavior is enabled by default, but can be turned of by
setting `config.exclude_searchroot = False`. (For path queries
of depth 0 or 1 it has no effect either way).

Omitting the searchroot from the results is achieved by
transforming the query:

A path query against a specific context will be transformed into a
multi-path query against all that context's immediate children
(which are determined beforehand via a second query).

This will cause the transformed query to return the exact same results,
except that the context queried against (the searchroot) now isn't
included in the results.